Create back button and correct background color
