+++ package.json
@@ -7,7 +7,8 @@
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-particles-js": "^1.5.0",
-    "react-router-dom": "^4.0.0"
+    "react-router-dom": "^4.0.0",
+    "rgb-hex": "^2.1.0"
   },
   "devDependencies": {
     "react-scripts": "0.9.5"
