+++ src/index.js
@@ -9,20 +9,12 @@ import HomePage from './components/HomePage';
 import PoemAnalysis from './components/PoemAnalysis';
 import PoemList from './components/PoemList';
 
+import { determineBackgroundColor } from './scripts/backgroundColor';
+
 class Index extends Component {
   // Select a semi-random background color for the page
   componentWillMount() {
-    document.body.style.backgroundColor = this.determineBackgroundColor();
-  }
-
-  determineBackgroundColor() {
-    const backgroundColors = [
-      '#1976D2', // blue
-      '#EF6C00', // orange
-      '#7E57C2', // purple
-      '#00897B' // teal
-    ];
-    return backgroundColors[Math.floor(Math.random() * backgroundColors.length)];
+    determineBackgroundColor()
   }
 
   // Render router component
