+++ src/styles/poemanalysis.css
@@ -74,6 +74,7 @@
   color: #fff;
   font-family: 'Open Sans';
   padding: 0.8em;
+  font-size: 1em;
   position: fixed;
   right: 0;
   top: 0;
@@ -96,4 +97,15 @@
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
+}
+
+.close-button {
+  color: #fff;
+  cursor: pointer;
+  font-size: 2em;
+  position: fixed;
+  top: 0;
+  left: 0;
+  padding: 0.4em;
+  z-index: 11;
 }
